Movido 'Pending' para um arquivo separado
-- Movido schema e interface pending para um mesmo arquivo 'pending.ts'
