+++ src/db/models/person/currencies/interface.ts
@@ -1,21 +1,5 @@
-import { Decimal128, ObjectId } from 'mongodb'
-
-/** A interface de operações de alteração de saldo pendentes*/
-export interface Pending {
-	/**
-	 * Referencia ao objectId da operação em sua respectiva collection
-	 */
-	opid: ObjectId
-	/**
-	 * O tipo da operação, para identificar em qual collection ela está
-	 */
-	type: 'transaction'
-	/**
-	 * O amount da operação. Positivo se é uma operação que aumenta o saldo do
-	 * usuário e negativo caso seja uma operação que reduzirá seu saldo
-	 */
-	amount: Decimal128
-}
+import { Decimal128 } from 'mongodb'
+import { Pending } from './pending'
 
 /** A interface de uma currency da colletion people */
 interface Currency {
